move context to productReviewContext
